I'm fed up of new labs users having problems on initial login because their key was in the wrong format.
When the key doesn't look to be in the right format:
a) Try to automatically convert it in the server. Much easier
than explaining where to find the option in the menu.
b) Reject it in other case, so at least they know upfront that, 
and don't have to discover it the hard way when trying to login
after discarding the other possible reasons.