Check the status of Nova requests, don't just assume they succeed
Instance list requests seem to fail about 50% of the time. Probably
other requests regularly fail as well. It's better to abort the request
and avoid possible data corruption than to continue with an "invalid
value supplied for foreach" warning. Throwing an exception also makes
it easier to track and fix the root cause.

Change-Id: I6b4a255b69e3877d155c3f016b801c943ec6e467