Fix for OpenStack token loss on memcache restart
Fixes bug 42596. The token was being loaded from the database too
late in the authentication chain, which showed that people had
credentials, but didn't properly use them against keystone. This
change moved the loading of the generic token to the spot where
we detect that the token is missing in memcache. When the token
is loaded from the database, it's then stuck back into memcache.

Change-Id: I28a2847193df2a36f66c34b43425f4c4df923359