Call $project->editArticle for all member changes.
Previously we called this from the GUI when a
member was explicitly added or removed; this
missed the case where memebers were added
to bastion thanks to a change in shell rights.

So, now the call to editArticle() is in the actual
add/delete member functions inside the project class
so we should catch every single one.

Change-Id: Id9fc29747a0838f6c208cdc197a0273723e92ee3