Don't record members on bastion project page
It's too slow and causes failures when giving new users shell access
(by adding them to other projects).
Given that wgOpenStackManagerBastionProjectName signifies a project
that all users are to be added to, I think it's reasonable that we don't
bother providing a list of members.
If necessary, it's still possible to access this information via LDAP in labs,
with ldaplist -l group project-bastion
(This has been tested in production)

Bug: T114229
Change-Id: I945d9e752ccc7ebb0c93ef11258c7b7479a2aa80