Inject $wgAuth = LdapAuthenticationPlugin on WMF CI
MediaWiki now has a structure test that check for API documentation
validity. To inspect the API doc, the test invokes our API entry point
which ends up relying on LdapAuthentication plugin.

Abuse $wgWikimediaJenkinsCI to set $wgAuth to LdapAuthenticationPlugin.
This is done as an extension function which is loaded rather late in the
setup.

Bug: T124613
Change-Id: Ifee416b7a9c5837e6141df9da1dc46b2726e67fd