Get projects and membership via the keystone api
Note that this does not /write/ via the keystone api --
that will be done in a future patch

This patch can't really be rolled out on its own,
since keystone won't understand about project users until
the 'users' role is created and populated (which will never
happen in ldap).  Also, keystone tokens
and memcached should be purged when the whole patchset
is rolled out.

Depends-On: Ic8e9456be2170885fa6d7fa435b691a2ef47a981
Bug: T115029
Change-Id: Ibd728d72260c583ef2fd1ec0fca8e2c583e091f7