Abolish wgOpenStackManagerLDAPUseUidAsNamingAttribute
This has been set to 'true' for eternity.  The 'false' condition
was, in places, unimplemented, and would have produced inconsisten
results in any case.

After the associated puppet patch, keystone will use the actual
username as the username field, so want OSM to get in line
and stop substituting in uids for usernames.

Change-Id: I634c1502463010dc7d53d23adddb04b13e1418c0
Depends-on: I5bcedc6dbed1fd0a97f8f9001f8a5c590a0fb764