Instance creation:  remove check for the 'show' metadata item
Glance now supports the 'deactivated' image status which has
the equivalent meaning to our home-made 'show' setting:  deactivated
images are still visible but cannot be used for new instances.

'deactivated' is supported by Horizon, so we should standardize
on that setting and stop using 'show'.

Conveniently, we already had a test for 'status' so the check
for 'show' is now redundant.

Documented here:

https://wikitech.wikimedia.org/w/index.php?title=OpenStack&type=revision&diff=349377&oldid=294218

Change-Id: Id62f3863d74e9b1596a4297131191f25bce3d1f1