ApiListNovaInstances: only update memc on cache miss
Move the memc->set() inside the cache miss handling block. With it
outside the miss block we are effectively caching forever as long as the
api is being polled more frequently than once every 60 seconds. This is
most likely to show up when someone is testing the api and will confuse
them greatly.

Change-Id: I33c67565c1611ee92ecdc2e7d47657d4be0ae7a3