Do not create sudo policies for chown ("-chmod")
Initially on Tool Labs users were supposed to fix ownership issues in
the home directories of their tools by executing:

| sudo /bin/chown -R tools.$TOOL:tools.$TOOL /data/project/$TOOL

This usage was never promoted and so did not catch on, but was
replaced by the utility take(1) which allows tool accounts to assume
ownership of files in their home directories if they share a group
with the files.

This change thus removes the creation of the unpromoted and unused
sudo policies.  After merging, existing sudo policies
"tools.$TOOL-chmod" can be removed manually.

Change-Id: Ie13f33765e7c3995b001e754ed2c8e81eb1eea3a