Split ext.openstack.base JavaScript from ext.openstack style module
The 'ext.openstack' module is only ever loaded through addModuleStyles()
in this extension. Yet, it also contains JS code and dependencies
(which are ignored by addModuleStyles).

The other two modules (..Address and ..Instance) depended on this
'ext.openstack' pseudo-style module, which would then cause the
styles to load for a second time, and this time with the JS code
and dependencies as well.

Split off this re-usable JS portion into its own module.

Bug: T92459
Change-Id: Ia430fa103b7d2524fdc8ddbe9a0b73481d8f7137